pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "pythonflask"
version = "0.1.0"
[tool.poetry.dependencies]
Flask = "^1.1"
MLB-StatsAPI = "^0.1.7"
WTForms = "^2.3"
flask-wtf = "^0.14.3"
python = "^3.8"
[tool.poetry.dev-dependencies]
